migrated attack code to basenode
